+++ src/main/java/com/example/exoevalcabinet/repositories/InfirmiereRepository.java
@@ -3,5 +3,5 @@ package com.example.exoevalcabinet.repositories;
 import com.example.exoevalcabinet.models.Infirmiere;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InfirmiereRepository extends JpaRepository<Infirmiere,Long> {
+public interface InfirmiereRepository extends JpaRepository<Infirmiere,Integer> {
 }
